+++ packages/user-taxonomies/src/fields/visibility.tsx
@@ -1,52 +1,19 @@
 /**
  * WordPress dependencies
  */
-import type { Field, Form } from '@wordpress/dataviews';
+import type { Form } from '@wordpress/dataviews';
 import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
-import type { TaxonomyFormData } from '../types';
+import { booleanField } from './utils';
 
-type VisibilityKey =
-	| 'publicly_queryable'
-	| 'show_ui'
-	| 'show_in_menu'
-	| 'show_in_nav_menus'
-	| 'show_tagcloud'
-	| 'show_in_quick_edit'
-	| 'show_admin_column'
-	| 'show_in_rest';
-
-type BooleanOptions = {
-	description: string;
-	isVisible?: ( item: TaxonomyFormData ) => boolean;
-};
-
-function booleanField(
-	id: VisibilityKey,
-	label: string,
-	options: BooleanOptions
-): Field< TaxonomyFormData > {
-	const field: Field< TaxonomyFormData > = {
-		id,
-		label,
-		type: 'boolean',
-		description: options.description,
-		Edit: 'toggle',
-		getValue: ( { item } ) => item.config[ id ],
-		setValue: ( { item, value } ) => ( {
-			config: { ...item.config, [ id ]: !! value },
-		} ),
-		filterBy: false,
-		enableSorting: false,
-	};
-	if ( options.isVisible ) {
-		field.isVisible = options.isVisible;
-	}
-	return field;
-}
+export const publicField = booleanField( 'public', __( 'Public' ), {
+	description: __(
+		'Whether a taxonomy is intended for use publicly either via the admin interface or by front-end users.'
+	),
+} );
 
 export const showInRestField = booleanField(
 	'show_in_rest',
@@ -129,6 +96,7 @@ export const showTagcloudField = booleanField(
 );
 
 export const visibilityFormFields: Form[ 'fields' ] = [
+	'public',
 	'show_in_rest',
 	'publicly_queryable',
 	'show_ui',
